Moves all of the HAT claim logic into Authentication controller
